Outsource createMatchesQuery
